# Conflicts:
#	viewer/CMakeLists.txt
#	viewer/content/CylindersFromFile.cpp

+++ hayStack/Cylinders.cpp
@@ -21,7 +21,7 @@ namespace hs {
   box3f Cylinders::getBounds() const
   {
     box3f bounds;
-    int   numCylinders = indices.empty() ? (vertices.size()/2) : indices.size();
+    int   numCylinders = indices.empty() ? ((int)vertices.size()/2) : (int)indices.size();
     for (int i=0;i<numCylinders;i++) {
       vec2i idx = indices.empty() ? (vec2i(2*i)+vec2i(0,1)) : indices[i];
       float r = radii.empty()?radius:radii[i];
